Adicionado controle de exceções.

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.alura.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -66,23 +67,37 @@ public class TopicosController {
 
 	// http://localhost:8080/topicos/12
 	@GetMapping("/{id}") // isso aqui é uma url dinamica que recebe um parametro com o nome id
-	public DetalheTopicoDto detalhar(@PathVariable Long id) {
+	public ResponseEntity detalhar(@PathVariable Long id) {
 		// @PathVariable indica que o parametro vem pela URI e não por ?
-		Topico topico = topicoRepository.getOne(id);
-		return new DetalheTopicoDto(topico);
+		Optional<Topico> topico = topicoRepository.findById(id);
+		if (topico.isPresent()) {
+			return ResponseEntity.ok(new DetalheTopicoDto(topico.get()));
+		}
+
+		return ResponseEntity.notFound().build();
 	}
 
 	@PutMapping("/{id}")
 	@Transactional
 	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualiacaoTopicoForm params) {
-		Topico topico = params.atualizar(id, topicoRepository);
-		return ResponseEntity.ok(new TopicoDto(topico));
+		Optional<Topico> topico = topicoRepository.findById(id);
+		if (topico.isPresent()) {
+			Topico topicoAtualizado = params.atualizar(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDto(topicoAtualizado));
+		}
+
+		return ResponseEntity.notFound().build();
 	}
 
 	@DeleteMapping("/{id}")
 	@Transactional
 	public ResponseEntity<?> remover(@PathVariable Long id) {
-		topicoRepository.deleteById(id);
-		return ResponseEntity.ok().build();
+		Optional<Topico> topico = topicoRepository.findById(id);
+		if (topico.isPresent()) {
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build();
+		}
+
+		return ResponseEntity.notFound().build();
 	}
 }
